git-svn-id: https://opensource.kantega.no/svn/openaksess/trunk@5163 fd808399-8219-4f14-9d4c-37719d9ec93d

+++ modules/solr-search/src/java/no/kantega/openaksess/search/solr/index/SolrDocumentIndexer.java
@@ -77,6 +77,9 @@ public class SolrDocumentIndexer implements DocumentIndexer {
     }
 
     public void deleteById(List<String> uids) {
+        if (uids.size() == 0) {
+            return;
+        }
         try {
             UpdateResponse updateResponse = solrServer.deleteById(uids, 0);
         } catch (Exception e) {
